backend/requirements.txt
# Google ADK and AI
google-adk>=1.1.0
google-cloud-aiplatform>=1.60.0
google-cloud-bigquery>=3.25.0

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# WebSocket support
websockets>=12.0

# Utilities
python-dateutil>=2.9.0
